+++ server/src/module-models/module-models.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { ModuleModelsController } from './module-models.controller';
+import { ModuleModelsService } from './module-models.service';
+
+@Module({
+  controllers: [ModuleModelsController],
+  providers: [ModuleModelsService],
+})
+export class ModuleModelsModule {}